# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel DDEV API Documentation",
        "description": "Swagger OpenApi description for our Laravel backend.",
        "contact": {
            "email": "developer@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://backend.ddev.site/api",
            "description": "Local Environment Server"
        }
    ],
    "paths": {
        "/api/healthcheck": {
            "get": {
                "description": "Healthcheck endpoint",
                "operationId": "healthcheck",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/todos": {
            "get": {
                "tags": [
                    "Todos"
                ],
                "summary": "Get list of tasks",
                "description": "Returns a list of all Todo items",
                "operationId": "getTodosList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Buy groceries"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Milk, bread, and eggs",
                                                "nullable": true
                                            },
                                            "is_completed": {
                                                "type": "boolean",
                                                "example": false
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Todos"
                ],
                "summary": "Create new Todo",
                "description": "Creates a new Todo item and returns it",
                "operationId": "storeTodo",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Buy groceries",
                                        "maxLength": 255
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Supermarket visit",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Buy groceries"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Supermarket visit",
                                            "nullable": true
                                        },
                                        "is_completed": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/todos/{todo}": {
            "get": {
                "tags": [
                    "Todos"
                ],
                "summary": "Get Todo details",
                "description": "Returns a single Todo item",
                "operationId": "getTodoById",
                "parameters": [
                    {
                        "name": "todo",
                        "in": "path",
                        "description": "Todo ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Buy groceries"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Supermarket visit",
                                            "nullable": true
                                        },
                                        "is_completed": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Todo not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Todos"
                ],
                "summary": "Update existing Todo",
                "description": "Updates an existing Todo item",
                "operationId": "updateTodo",
                "parameters": [
                    {
                        "name": "todo",
                        "in": "path",
                        "description": "Todo ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Buy groceries updated",
                                        "maxLength": 255
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated description",
                                        "nullable": true
                                    },
                                    "is_completed": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Buy groceries updated"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Updated description",
                                            "nullable": true
                                        },
                                        "is_completed": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Todo not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Todos"
                ],
                "summary": "Delete Todo",
                "description": "Deletes a specified Todo item",
                "operationId": "deleteTodo",
                "parameters": [
                    {
                        "name": "todo",
                        "in": "path",
                        "description": "Todo ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Todo deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Todo not found"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Todos",
            "description": "Todos"
        }
    ]
}